feat: preview file outputs directly
