+++ ui/src/components/logs/TaskRunDetails.vue
@@ -248,7 +248,7 @@
     import {useExecutionsStore} from "../../stores/executions"
     import TaskRunLine from "../executions/TaskRunLine.vue"
     import * as FlowUtils from "../../utils/flowUtils"
-    import FilePreview from "../executions/FilePreview.vue"
+    import FilePreview from "../executions/FilePreviewDrawer.vue"
     import {apiUrl} from "override/utils/route"
     import * as Utils from "../../utils/utils"
     import * as LogUtils from "../../utils/logs"
